Initial commit
A quick CLI tool looking at automatic conversion between different
config-ish data formats.

It relies on serde and its ecosystem for the conversion. Getting this
working was a little bit of a challenge since my prior experience with
Serde was in a much more static situation, where I was working with
structures known ahead of time, and with a fixed format. Using
erased_serde and judicious heap allocation I've gotten it to
at least work.

Before adding more supported formats it'd be nice to:

* document it
* find a way to tidy up the erased deserializer/serializer creation
* figure out if leaking objects will lead to stdin & stdout not being
  closed, and what the implications of that are. If they were
  disk-backed files I'd be a bit more worried